SytheticTableGenerator can't generate NULL columns
For the column calibration we wanted to explore how NULL values affect Scan performance.
During this we encountered that with the current version of
the Synthetic Table Generator it is not possible to generate these values.

Is this a bug?

```
terminate called after throwing an instance of 'std::logic_error'
  what():  src/lib/resolve_type.hpp:76 data_type cannot be null.

Process finished with exit code 134 (interrupted by signal 6: SIGABRT)
```